Fixed various bugs and added various options

+++ happyweb/admin/widgets/slideshow/delete.php
@@ -1,8 +1,10 @@
 <?php
 $files = json_decode($data->filenames);
-foreach($files as $obj) {
-  // delete the original images
-  unlink($_SERVER["DOCUMENT_ROOT"]."/your_site/uploaded_files/originals/".$obj->id);
-  // delete the resized images
-  unlink($_SERVER["DOCUMENT_ROOT"]."/your_site/uploaded_files/large/".$obj->id);  
+if ($data->filenames != "") {
+  foreach($files as $obj) {
+    // delete the original images
+    unlink($_SERVER["DOCUMENT_ROOT"]."/your_site/uploaded_files/originals/".$obj->id);
+    // delete the resized images
+    unlink($_SERVER["DOCUMENT_ROOT"]."/your_site/uploaded_files/large/".$obj->id);  
+  }
 }
